introducing DataToActuators
useful for PID, Bayesian ... extensions to set various actiions for actuators
